+++ app/src/main/java/com/example/android/ShotTracker/db/ShotLinkDAO.java
@@ -51,7 +51,7 @@ public class ShotLinkDAO extends ShotTrackerDBDAO {
 
         while (cursor.moveToNext()){
             ShotType shotType = new ShotType();
-            shotType.setID(cursor.getInt(0));
+            shotType.setID(cursor.getLong(0));
             shotType.setType(cursor.getString(1));
 
             shotTypes.add(shotType);
